requirements.txt
flask>=2.0.1
pillow>=9.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
requests>=2.25.1
lime-wire>=1.0.0